add changes from code review
