+++ src/snowflake/snowpark/modin/plugin/docstrings/window.py
@@ -507,7 +507,10 @@ class Rolling:
         other : Series or DataFrame, optional
             If not supplied then will default to self and produce pairwise output.
         pairwise : bool, default None
-            If False then only matching columns between self and other will be used and the output will be a DataFrame. If True then all pairwise combinations will be calculated and the output will be a MultiIndexed DataFrame in the case of DataFrame inputs. In the case of missing elements, only complete pairwise observations will be used.
+            If False then only matching columns between self and other will be used and the output will be a DataFrame.
+            If True then all pairwise combinations will be calculated and the output will be a MultiIndexed DataFrame
+            in the case of DataFrame inputs. In the case of missing elements, only complete pairwise observations
+            will be used.
         ddof : int, default 1
             Delta Degrees of Freedom. The divisor used in calculations is ``N - ddof``,
             where ``N`` represents the number of elements.
